add:  - docker-compose.yml testing and bug fixing

+++ src/main/java/com/example/dripchip/service/impl/LocationServiceImpl.java
@@ -3,8 +3,10 @@ package com.example.dripchip.service.impl;
 import com.example.dripchip.dto.LocationDTO.Response;
 import com.example.dripchip.dto.LocationDTO.Request;
 import com.example.dripchip.entites.LocationPoint;
+import com.example.dripchip.exception.BadRequestException;
 import com.example.dripchip.exception.ConflictException;
 import com.example.dripchip.exception.NotFoundException;
+import com.example.dripchip.repositorie.AnimalDAO;
 import com.example.dripchip.repositorie.LocationDAO;
 import com.example.dripchip.service.LocationService;
 import jakarta.validation.Valid;
@@ -13,6 +15,9 @@ import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -58,11 +63,14 @@ public class LocationServiceImpl implements LocationService {
         return parseToDTO(locationPoint);
     }
 
-
     @Override
-    public void deleteById(@Min(1) @NotNull Long pointId) throws NotFoundException {
+    public void deleteById(@Min(1) @NotNull Long pointId) throws NotFoundException, BadRequestException {
 
-        findLocationById(pointId);
+        LocationPoint locationPoint = findLocationById(pointId);
+
+        if (locationPoint.getVisitedLocations().size() > 0) {
+            throw new BadRequestException("Some animal associated with this location");
+        }
 
         locationDAO.deleteById(pointId);
     }
